requirements.txt
langgraph
langchain
langchain-openai
streamlit
python-dotenv
fastapi
uvicorn
pydantic>=2.7.4
openai
requests
mcp